Interpersonal Relationships Survey Data

Simulated survey data (n = 574) from college students at a predominantly white rural state university, containing demographics, relationship variables, and scores on several interpersonal and psychological scales including the DAQ, Interpersonal Dependency Inventory, Interpersonal Reactivity Index, Sociability Scale, RISC, GCB, and LSAS-SR.

Usage

interpersonal_data

Format

A tibble with 574 rows and 33 variables:

age

Age in years (18-25)

gender

Self-described gender: Male (1), Female (2), Another (3)

sexorient

Sexual orientation (1-11, see value labels)

race

Self-described race: Asian (1), Black (2), Indigenous/Aboriginal/First Nations (3), Latino/Hispanic (4), Middle Eastern (5), White (6), Other (7)

hand

Writing hand: Right (1), Left (2), Both (3)

community

Childhood community type: Rural (1), Small town (2), Suburban (3), Urban (4)

parentedu

Parent graduated from four-year college: No (0), Yes (1)

famclass

Childhood social class: Working class (1), Lower class (2), Lower middle class (3), Upper middle class (4), Upper class (5)

faminc

Family income during senior year of high school

numsib

Number of siblings

move

Number of times moved as a child

clsfrn

Number of close friends

clsfrlst

Number of people who would list you as a close friend

greek_in

Fraternity/sorority membership: Independent (1), Greek (2)

campus

Living on campus: No (0), Yes (1)

relsp

Romantic relationship status (1-5, see value labels)

rlength

Length of current or last relationship in months

serious

Seriousness rating of current/most recent relationship (1-7)

numrels

Number of dating relationships

gcb

Generic Conspiracist Beliefs scale score

datdaq

Dating subscale of the DAQ

assrtdaq

Assertiveness subscale of the DAQ

emorel

Emotional Reliance on Others subscale of the IDI

lacksc

Lack of Self-Confidence subscale of the IDI

auto

Assertion of Autonomy subscale of the IDI

perspec

Perspective-Taking subscale of the IRI

fantasy

Fantasy subscale of the IRI

empath

Empathic Concern subscale of the IRI

distress

Personal Distress subscale of the IRI

polsoc

Political Sociability subscale of Sociability Scale

npolsoc

Non-Political Sociability subscale of Sociability Scale

risc

Relational-Interdependent Self-Construal Scale score

lsas

Liebowitz Social Anxiety Scale self-report total (LSAS-SR)

Source

Simulated data generated to resemble plausible survey responses from undergraduate psychology students.